fix son DB, in users: user_id - bigint

+++ dbscripts/son_dispatcher.sql
@@ -18,7 +18,7 @@ CREATE TABLE users
 (
   id INT(4) NOT NULL AUTO_INCREMENT,
     org_id INT(4) NOT NULL,
-    user_id int(25) NOT NULL unique,
+    user_id bigint NOT NULL unique,
     user_name varchar(32) not null,
     CONSTRAINT PK_users
     PRIMARY KEY(ID),
